Rename actions modules for consistency
