+++ kolibri/plugins/management/assets/src/device_management/state/actions/manageContentActions.js
